requirements.txt
gspread
pandas
flask
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
sqlalchemy
pymongo
mysql-connector-python
git+https://github.com/bjahnke/seatgeek_pandas_wrapper.git